backend/requirements.txt
fastapi
uvicorn
python-multipart
aiohttp
requests
python-dotenv
volcengine
tos
